Move client into submodule
